backend_service/requirements.txt
# Backend Service Dependencies
# Main application dependencies
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
email-validator==2.1.0
pydantic[email]==2.5.0

# Testing dependencies
pytest==7.4.3
requests==2.31.0
